test: reuse setup code
