+++ test/runner/test_events.py
@@ -1,14 +1,8 @@
-import io
-from datetime import timedelta
-
 import pytest
-import requests
-from urllib3 import HTTPResponse
 
-from schemathesis.models import Case, CaseSource, Check, Status
+from schemathesis.models import CaseSource, Check, Status
 from schemathesis.runner import events
 from schemathesis.runner.serialization import SerializedCheck
-from schemathesis.utils import WSGIResponse
 
 
 def test_unknown_exception():
@@ -20,38 +14,14 @@ def test_unknown_exception():
         assert event.exception.strip() == "ZeroDivisionError: division by zero"
 
 
-@pytest.fixture
-def case_factory(swagger_20):
-    def factory():
-        return Case(operation=swagger_20["/users"]["GET"])
-
-    return factory
-
-
-@pytest.fixture(params=[requests.Response, WSGIResponse])
-def response_factory(request, mocker):
-    def factory(headers):
-        response = mocker.create_autospec(request.param)
-        response.status_code = 500
-        response.reason = "Internal Server Error"
-        response.encoding = "utf-8"
-        response.elapsed = timedelta(1.0)
-        response.headers = {}
-        response.response = []
-        response.raw = HTTPResponse(body=io.BytesIO(b""), status=500, headers={})
-        response.request = requests.PreparedRequest()
-        response.request.prepare(method="POST", url="http://127.0.0.1", headers=headers)
-        return response
-
-    return factory
-
-
-def test_serialize_history(case_factory, response_factory):
+@pytest.mark.parametrize("factory_name", ("requests", "werkzeug"))
+def test_serialize_history(case_factory, response_factory, factory_name):
+    factory = getattr(response_factory, factory_name)
     root_case = case_factory()
     value = "A"
-    root_case.source = CaseSource(case=case_factory(), response=response_factory({"X-Example": value}), elapsed=1.0)
+    root_case.source = CaseSource(case=case_factory(), response=factory(headers={"X-Example": value}), elapsed=1.0)
     check = Check(
-        name="test", value=Status.failure, response=response_factory({"X-Example": "B"}), elapsed=1.0, example=root_case
+        name="test", value=Status.failure, response=factory(headers={"X-Example": "B"}), elapsed=1.0, example=root_case
     )
     serialized = SerializedCheck.from_check(check)
     assert len(serialized.history) == 1
